+++ packages/colorpickers/src/elements/ColorpickerDialog/index.tsx
@@ -8,6 +8,7 @@
 import React, {
   useState,
   useRef,
+  useEffect,
   Children,
   cloneElement,
   forwardRef,
@@ -28,6 +29,10 @@ import {
   StyledTooltipModal,
   StyledTooltipBody
 } from '../../styled';
+
+interface IDialogChanges {
+  isOpen?: boolean;
+}
 
 export interface IColorpickerDialogProps extends IColorpickerProps {
   /**
@@ -57,6 +62,10 @@ export interface IColorpickerDialogProps extends IColorpickerProps {
    */
   isAnimated?: boolean;
   /**
+   * Opens the dialog in a controlled color picker dialog
+   */
+  isOpen?: boolean;
+  /**
    * Applies inset `box-shadow` styling on focus
    */
   focusInset?: boolean;
@@ -64,6 +73,12 @@ export interface IColorpickerDialogProps extends IColorpickerProps {
    * Passes HTML attributes to the color dialog button element
    */
   buttonProps?: HTMLAttributes<HTMLButtonElement>;
+  /**
+   * Handles dialog changes
+   *
+   * @param {Object} changes The changed dialog state
+   */
+  onDialogChange?: (changes: IDialogChanges) => void;
 }
 
 /**
@@ -84,17 +99,20 @@ export const ColorpickerDialog = forwardRef<
       hasArrow,
       isAnimated,
       isOpaque,
+      isOpen,
       popperModifiers,
       zIndex,
       focusInset,
       disabled,
       buttonProps,
+      onDialogChange,
       children,
       ...props
     },
     ref
   ) => {
     const isControlled = color !== null && color !== undefined;
+    const isDialogControlled = isOpen !== undefined && isOpen !== null;
     const buttonRef = useRef<HTMLButtonElement>(null);
     const colorPickerRef = useRef<HTMLDivElement>(null);
     const [referenceElement, setReferenceElement] = useState<HTMLButtonElement | null>();
@@ -102,13 +120,33 @@ export const ColorpickerDialog = forwardRef<
       defaultColor
     );
 
+    const openDialog = () => {
+      setReferenceElement(buttonRef.current);
+      onDialogChange && onDialogChange({ isOpen: true });
+    };
+
+    const closeDialog = () => {
+      setReferenceElement(null);
+      onDialogChange && onDialogChange({ isOpen: false });
+    };
+
     const onClick = composeEventHandlers(props.onClick, () => {
       if (referenceElement) {
-        setReferenceElement(null);
+        closeDialog();
       } else {
-        setReferenceElement(buttonRef.current);
+        openDialog();
       }
     });
+
+    useEffect(() => {
+      if (isDialogControlled) {
+        if (isOpen) {
+          setReferenceElement(buttonRef.current);
+        } else {
+          setReferenceElement(null);
+        }
+      }
+    }, [isOpen, isDialogControlled]);
 
     return (
       <>
@@ -143,7 +181,7 @@ export const ColorpickerDialog = forwardRef<
           placement={placement}
           referenceElement={referenceElement}
           onClose={() => {
-            setReferenceElement(null);
+            closeDialog();
             onClose && onClose(isControlled ? (color as IColor) : (uncontrolledColor as IColor));
           }}
           {...props}
@@ -183,6 +221,7 @@ ColorpickerDialog.propTypes = {
   ]),
   onClose: PropTypes.func,
   onChange: PropTypes.func,
+  onDialogChange: PropTypes.func,
   disabled: PropTypes.bool,
   labels: PropTypes.object,
   color: PropTypes.oneOfType<any>([PropTypes.object, PropTypes.string]),