+++ packages/colorpickers/stories/examples/ColorpickerDialog/WithFormInput.tsx
@@ -8,6 +8,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { Story, Meta } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
 import { Field, Label, Input, InputGroup } from '@zendeskgarden/react-forms';
 import { IColor, ColorpickerDialog } from '@zendeskgarden/react-colorpickers';
 import { Col, Grid, Row } from '@zendeskgarden/react-grid';
@@ -40,7 +41,8 @@ export const WithFormInput: Story = ({
   alpha,
   disabled,
   hasArrow,
-  isAnimated
+  isAnimated,
+  isOpen
 }) => {
   const labels = { alphaSlider, hueSlider, hex, red, green, blue, alpha };
   const [input, setInput] = useState('#17494d99');
@@ -103,6 +105,8 @@ export const WithFormInput: Story = ({
                 disabled={disabled}
                 hasArrow={hasArrow}
                 isAnimated={isAnimated}
+                isOpen={isOpen}
+                onDialogChange={action('onDialogChange')}
               />
             </InputGroup>
           </StyledField>
@@ -123,6 +127,7 @@ WithFormInput.args = {
   disabled: false,
   hasArrow: false,
   isAnimated: true,
+  isOpen: false,
   placement: 'bottom-end'
 };